feat:allow queryservice be hosted at non-root path
The RequestsSessionWithEndpoint builds a URI from a base URL + a path.

Currently in all cases of the RequestsSessionWithEndpoint being used in
the QueryServiceClient, the paths that are being passed in are all
absolute paths.

By passing in absolute paths, it makes it impossible to use a base URL
that also includes a path (e.g. http://localhost:8000/qs/), because
urllib.parse.urljoin will create the new full URI by substituting the
path of the base URL (e.g. /qs/) with the full absolute path that is
passed in to the RequestsSessionWithEndpoint constructor.

By passing in relative paths to the RequestsSessionWithEndpoint
constructor, the ability to host the query service on a base URL with a
path is now possible.